refactor(container): add AppMenuColumn.tsx
