+++ package.json
@@ -82,9 +82,9 @@
     "@types/react-color": "^2.11.2",
     "@types/react-dom": "^15.5.0",
     "@types/url-join": "^0.8.2",
-    "almin": "0.13.0",
+    "almin": "0.13.2",
     "almin-logger": "^4.0.3",
-    "almin-react-container": "0.3.0",
+    "almin-react-container": "0.3.2",
     "classnames": "^2.2.5",
     "combokeys": "^2.4.6",
     "debug": "^2.6.8",
