+++ src/component/container/AppContainer.tsx
@@ -2,7 +2,6 @@
 "use strict";
 import * as React from "react";
 import { AppStoreGroupState } from "../../store/AppStoreGroup";
-import { GitHubSearchContainer } from "./GitHubSearchContainer/GitHubSearchContainer";
 import { GitHubSearchStreamContainer } from "./GitHubSearchStreamContainer/GitHubSearchStreamContainer";
 import { BaseContainer } from "./BaseContainer";
 import IframeBrowser from "../project/IframeBrowser/IframeBrowser";
@@ -10,6 +9,7 @@ import { ShortcutKeyContainer } from "./ShortcutKeyContainer/ShortcutKeyContaine
 import { QuickIssueContainer } from "./QuickIssueContainer/QuickIssueContainer";
 import { QuerySettingContainer } from "./QuerySettingContainer/QuerySettionContainer";
 import { GitHubSettingPanelContainer } from "./GitHubSettingPanelContainer/GitHubSettingPanelContainer";
+import { AppMenuColumn } from "./AppMenuColumn/AppMenuColumn";
 
 export class AppContainer extends BaseContainer<AppStoreGroupState, {}> {
     render() {
@@ -21,12 +21,11 @@ export class AppContainer extends BaseContainer<AppStoreGroupState, {}> {
         return (
             <div className="AppContainer">
                 <ShortcutKeyContainer />
-                <nav className="AppContainer-nav">
-                    <GitHubSearchContainer
-                        gitHubSetting={this.props.gitHubSetting}
-                        gitHubSearchList={this.props.gitHubSearchList}
-                    />
-                </nav>
+                <AppMenuColumn
+                    className="AppContainer-nav"
+                    gitHubSetting={this.props.gitHubSetting}
+                    gitHubSearchList={this.props.gitHubSearchList}
+                />
                 <main className="AppContainer-main">
                     <GitHubSearchStreamContainer
                         app={this.props.app}
