+++ src/component/container/AppMenuColumn/AppMenuHeaderContainer/AppMenuHeaderContainer.css
@@ -1,0 +1,2 @@
+.AppMenuHeaderContainer {
+}
