+++ src/component/container/AppMenuColumn/GitHubSearchContainer/GitHubSearchContainer.css
@@ -1,0 +1,3 @@
+.GitHubSearchContainer {
+    padding-left: 1rem;
+}
