+++ src/component/container/AppMenuColumn/GitHubSettingContainer/GitHubSettingContainer.css
@@ -1,0 +1,3 @@
+.GitHubSettingContainer {
+    padding-left: 1rem;
+}
