--- src/component/container/GitHubSearchContainer/GitHubSearchContainer.css
@@ -1,3 +1,0 @@
-.GitHubSearchContainer-main {
-    padding-left: 1rem;
-}
